third_party/blink/web_tests/http/tests/security/media-element-audio-source-node-cross-origin-expected.txt

CONSOLE WARNING: The ScriptProcessorNode is deprecated. Use AudioWorkletNode instead. (https://bit.ly/audio-worklet)
CONSOLE MESSAGE: MediaElementAudioSource outputs zeroes due to CORS access restrictions for http://localhost:8000/security/resources/webaudio/laughter.wav
MediaElementAudioSourceNode with cross-origin source

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS All samples correctly zeroed.
PASS successfullyParsed is true

TEST COMPLETE
